MAINT!: rename "polarization" to "polarimetry"
