+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = polarization-sensitivity
-description = Symbolic polarization sensitivity
+name = polarimetry
+description = Symbolic expressions that describe an aligned polarimeter vector field
 author = Mikhail Mikhasenko, Remco de Boer
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -68,6 +68,6 @@ dev =
 where = src
 
 [options.package_data]
-polarization.lhcb =
+polarimetry.lhcb =
     model-definitions.yaml
     particle-definitions.yaml
