+++ src/polarimetry/function.py
