+++ src/polarimetry/lhcb/model-definitions.yaml
