+++ src/polarimetry/lhcb/particle-definitions.yaml
