+++ src/polarimetry/plot.py
