+++ src/polarimetry/spin.py
