+++ tests/test_decay.py
@@ -1,4 +1,4 @@
-from polarization.decay import IsobarNode, Particle
+from polarimetry.decay import IsobarNode, Particle
 
 # https://compwa-org--129.org.readthedocs.build/report/018.html#resonances-and-ls-scheme
 dummy_args = dict(mass=0, width=0)
